Changes to support Room Room 2.5.0-alpha03 / API 33 / 8.0.0-alpha03

+++ library/src/main/java/org/dbtools/android/room/CloseableDatabaseWrapper.kt
@@ -53,9 +53,9 @@ abstract class CloseableDatabaseWrapper<out T: RoomDatabase>(protected val conte
         try {
             val database = _database.get()
             database?.let {
-                val path = it.getOpenHelper().writableDatabase.path
+                val path = it.openHelper.writableDatabase.path
                 it.close()
-                if (deleteFile) {
+                if (deleteFile && path != null) {
                     DatabaseUtil.deleteDatabaseFiles(File(path))
                 }
             }
